setNeedsLayout when originally unknown size is set.
If the backing TTPhoto returned an unknown size (CGSizeZero), the photo
size is set by the TTPhotoView once an image is loaded from the network.
When this occurs, it is likely that the assumed size set by the
TTScrollView (the page size) is wrong, resulting in images overflowing
their page bounds. To remedy this, the TTPhotoView should be relaid out
once the image size is determined.