Allow thumbSize and thumbOrigin to be configured.
Using these properties, cells can be configured to
be used in different table view styles. For
example, a cell could be used in a grouped table
style by configuring the thumbSize = 67 and the
thumbOrigin to (10, 10) for the first row and
(10, 2) for every other row. Table rows should
also be adjusted to get the desired appearance.